requirements.txt
flask==3.0.2
python-dotenv==1.0.1
vertexai
google-cloud-aiplatform
waitress==3.0.0
pytest==8.0.0
google-generativeai
google-api-python-client
